+++ aldrin/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `Sender`, `UnboundReceiver`, `UnclaimedReceiver`, `PendingReceiver` and `Receiver`.
 - Add `UnboundSender::new()` and `UnboundReceiver::new()` to create channels directly from
   `ChannelCookie`s.
+- Add `client()` getters to `Sender` and `Receiver`.
 
 ## [0.6.0] - 2024-06-07